fix(web): restores error-feedback handling
