+++ web/src/engine/namespaced-main/dom/domManager.ts
@@ -14,20 +14,7 @@
 /// <reference path="variableStoreCookieSerializer.ts" />
 
 namespace com.keyman.dom {
-  // Utility object used to handle beep (keyboard error response) operations.
-  class BeepData {
-    e: HTMLElement;
-    c: string;
 
-    constructor(e: HTMLElement) {
-      this.e = e;
-      this.c = e.style.backgroundColor;
-    }
-
-    reset(): void {
-      this.e.style.backgroundColor = this.c;
-    }
-  }
 
   /**
    * This class serves as the intermediary between KeymanWeb and any given web page's elements.
@@ -44,10 +31,6 @@ namespace com.keyman.dom {
      * Implements stubs for the AliasElementHandlers interface for non-touch interaction.
      */
     nonTouchHandlers: DOMEventHandlers;
-
-    _BeepObjects: BeepData[] = [];  // BeepObjects - maintains a list of active 'beep' visual feedback elements
-    _BeepTimeout: number = 0;       // BeepTimeout - a flag indicating if there is an active 'beep'.
-                                    // Set to 1 if there is an active 'beep', otherwise leave as '0'.
 
     // Used for special touch-based page interactions re: element activation on touch devices.
     deactivateOnScroll: boolean = false;
@@ -92,71 +75,6 @@ namespace com.keyman.dom {
         console.error("Error occurred during shutdown");
         console.error(e);
       }
-    }
-
-    /**
-     * Function     beep          KB      (DOM-side implementation)
-     * Scope        Public
-     * @param       {Object}      Pelem     element to flash
-     * Description  Flash body as substitute for audible beep; notify embedded device to vibrate
-     */
-    doBeep(outputTarget: targets.OutputTarget) {
-      // Handles embedded-mode beeps.
-      let keyman = com.keyman.singleton;
-      if ('beepKeyboard' in keyman) {
-        keyman['beepKeyboard']();
-        return;
-      }
-
-      if(!(outputTarget instanceof targets.OutputTarget)) {
-        return;
-      }
-
-      // All code after this point is DOM-based, triggered by the beep.
-      var Pelem: HTMLElement = outputTarget.getElement();
-      if(outputTarget instanceof dom.targets.DesignIFrame) {
-        Pelem = outputTarget.docRoot; // I1446 - beep sometimes fails to flash when using OSK and rich control
-      }
-
-      if(!Pelem) {
-        return; // There's no way to signal a 'beep' to null, so just cut everything short.
-      }
-
-      if(!Pelem.style || typeof(Pelem.style.backgroundColor)=='undefined') {
-        return;
-      }
-
-      for(var Lbo=0; Lbo<this._BeepObjects.length; Lbo++) { // I1446 - beep sometimes fails to return background color to normal
-                                                                  // I1511 - array prototype extended
-        if(this._BeepObjects[Lbo].e == Pelem) {
-          return;
-        }
-      }
-
-      this._BeepObjects = com.keyman.singleton._push(this._BeepObjects, new BeepData(Pelem));
-      // TODO:  This is probably a bad color choice if "dark mode" is enabled.  A proper implementation
-      //        would probably require some 'fun' CSS work, though.
-      Pelem.style.backgroundColor = '#000000';
-      if(this._BeepTimeout == 0) {
-        this._BeepTimeout = 1;
-        window.setTimeout(this.beepReset.bind(this), 50);
-      }
-    }
-
-    /**
-     * Function     beepReset
-     * Scope        Public
-     * Description  Reset/terminate beep or flash (not currently used: Aug 2011)
-     */
-    beepReset(): void {
-      com.keyman.singleton.core.keyboardInterface.resetContextCache();
-
-      var Lbo;
-      this._BeepTimeout = 0;
-      for(Lbo=0;Lbo<this._BeepObjects.length;Lbo++) { // I1511 - array prototype extended
-        this._BeepObjects[Lbo].reset();
-      }
-      this._BeepObjects = [];
     }
 
     /* ------ Defines independent, per-control keyboard setting behavior for the API. ------ */
